refactoring and remove some files
refactoring and remove some files
