--- src/main/java/org/folio/marccat/integration/CatalogingHelper.java
@@ -1,201 +1,0 @@
-package org.folio.marccat.integration;
-
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import org.folio.marccat.exception.DataAccessException;
-import org.folio.marccat.exception.SystemInternalFailureException;
-import org.folio.marccat.exception.UnableToCreateOrUpdateEntityException;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-
-import javax.sql.DataSource;
-import java.sql.Connection;
-import java.sql.SQLException;
-import java.util.*;
-import java.util.function.BooleanSupplier;
-import java.util.function.Function;
-import java.util.stream.StreamSupport;
-
-import static java.util.stream.Collectors.toMap;
-import static org.folio.marccat.config.Global.HCONFIGURATION;
-
-/**
- * Helper functions used within the cataloging module.
- * Specifically, this class was originally thought as a supertype layer of each resource implementor; later, it has
- * been converted in this way (i.e. a collection of static methods) because https://issues.folio.org/browse/RMB-95
- *
- * @author agazzarini
- * @since 1.0
- */
-public abstract class CatalogingHelper {
-  private final static Properties DEFAULT_VALUES = new Properties();
-  private final static Map<String, DataSource> DATASOURCES = new HashMap<>();
-
-  static {
-    try {
-      DEFAULT_VALUES.load(CatalogingHelper.class.getResourceAsStream("/defaults.properties"));
-    } catch (final Throwable exception) {
-      throw new ExceptionInInitializerError(exception);
-    }
-  }
-
-  /**
-   * Executes a GET request.
-   *
-   * @param adapter           the bridge that carries on the existing logic.
-   * @param tenant            the tenant associated with the current request.
-   * @param configurator      the configuration client.
-   * @param configurationSets the requested configuration attributes sets.
-   */
-  public static <T> T doGet(
-    final PieceOfExistingLogicAdapter<T> adapter,
-    final String tenant,
-    final Configuration configurator,
-    final String... configurationSets) {
-    return exec(adapter, tenant, configurator, configurationSets);
-  }
-
-  /**
-   * Executes a POST request.
-   *
-   * @param adapter           the bridge that carries on the existing logic.
-   * @param tenant            the tenant associated with the current request.
-   * @param configurator      the configuration client.
-   * @param validator         a validator function for the entity associated with this resource.
-   * @param configurationSets the requested configuration attributes sets.
-   */
-  public static <T> ResponseEntity<T> doPost(
-    final PieceOfExistingLogicAdapter<T> adapter,
-    final String tenant,
-    final Configuration configurator,
-    final BooleanSupplier validator,
-    final String... configurationSets) {
-    if (validator.getAsBoolean()) {
-      final T result = exec(adapter, tenant, configurator, configurationSets);
-      final HttpHeaders headers = new HttpHeaders();
-      headers.add(HttpHeaders.CONTENT_TYPE, "application/json");
-      return new ResponseEntity<>(result, headers, HttpStatus.CREATED);
-    } else {
-      throw new UnableToCreateOrUpdateEntityException();
-    }
-  }
-
-  /**
-   * Executes a PUT request.
-   *
-   * @param adapter           the bridge that carries on the existing logic.
-   * @param tenant            the tenant associated with the current request.
-   * @param configurator      the configuration client.
-   * @param validator         a validator function for the entity associated with this resource.
-   * @param configurationSets the requested configuration attributes sets.
-   */
-  public static <T> void doPut(
-    final PieceOfExistingLogicAdapter<T> adapter,
-    final String tenant,
-    final Configuration configurator,
-    final BooleanSupplier validator,
-    final String... configurationSets) {
-    if (validator.getAsBoolean()) {
-      exec(adapter, tenant, configurator, configurationSets);
-    } else {
-      throw new UnableToCreateOrUpdateEntityException();
-    }
-  }
-
-  /**
-   * Executes a DELETE request.
-   *
-   * @param adapter           the bridge that carries on the existing logic.
-   * @param tenant            the tenant associated with the current request.
-   * @param configurator      the configuration client.
-   * @param configurationSets the requested configuration attributes sets.
-   */
-  public static <T> void doDelete(
-    final PieceOfExistingLogicAdapter<T> adapter,
-    final String tenant,
-    final Configuration configurator,
-    final String... configurationSets) {
-    exec(adapter, tenant, configurator, configurationSets);
-  }
-
-  /**
-   * Provides a unified approach (within the cataloging module) for wrapping an existing blocking flow.
-   *
-   * @param adapter           the bridge that carries on the existing logic.
-   * @param configurationSets the configurationSets required by the current service.
-   */
-  private static <T> T exec(
-    final PieceOfExistingLogicAdapter<T> adapter,
-    final String tenant,
-    final Configuration configurator,
-    final String... configurationSets) {
-    try {
-      final ObjectNode settings = configurator.attributes(tenant, true, configurationSets);
-      final DataSource datasource = datasource(tenant, settings);
-      try (final Connection connection = datasource.getConnection();
-           final StorageService service =
-             new StorageService(
-               HCONFIGURATION.buildSessionFactory().openSession(connection))) {
-        return adapter.execute(service, configuration(settings));
-      } catch (final SQLException exception) {
-        throw new DataAccessException(exception);
-      } catch (final Throwable exception) {
-        throw new SystemInternalFailureException(exception);
-      }
-    } catch (final Throwable throwable) {
-      throw new SystemInternalFailureException(throwable);
-    }
-  }
-
-  /**
-   * Creates a dedicated configuration for the current service.
-   *
-   * @param value the mod-configuration response.
-   * @return a dedicated configuration for the current service.
-   */
-  private static Map<String, String> configuration(final ObjectNode value) {
-    return StreamSupport.stream(value.withArray("configs").spliterator(), false)
-      .filter(node -> !"datasource".equals(node.get("configName").asText()))
-      .map(node -> new AbstractMap.SimpleEntry<>(node.get("code").asText(), node.get("value").asText()))
-      .collect(toMap(AbstractMap.SimpleEntry::getKey, AbstractMap.SimpleEntry::getValue));
-  }
-
-  /**
-   * Retrieves the datasource configuration from the given buffer.
-   * The incoming buffer is supposed to be the result of one or more calls to the mod-configuration module.
-   *
-   * @param value the configuration as it comes from the mod-configuration module.
-   * @return the datasource configuration used within this module.
-   */
-  private static DataSource datasource(final String tenant, final ObjectNode value) {
-    return DATASOURCES.computeIfAbsent(tenant, k -> newDataSourceInstance(value));
-  }
-
-  /**
-   * Creates a new datasource reference.
-   *
-   * @return a new datasource reference.
-   */
-  private static DataSource newDataSourceInstance(final ObjectNode value) {
-    final Map<String, String> config = StreamSupport.stream(value.withArray("configs").spliterator(), false)
-      .filter(node -> "datasource".equals(node.get("configName").asText()))
-      .map(node -> new AbstractMap.SimpleEntry<>(node.get("code").asText(), node.get("value").asText()))
-      .collect(toMap(AbstractMap.SimpleEntry::getKey, AbstractMap.SimpleEntry::getValue));
-    return DataSourceBuilder
-      .create()
-      .username(config.get("user"))
-      .password(config.get("password"))
-      .url("jdbc:postgresql://192.168.0.158:5433/folio_marccat_sv2")
-      .build();
-  }
-
-  /**
-   * A simple definition of a validation interface.
-   *
-   * @param <T> the kind of object that needs to be validated.
-   */
-  interface Valid<T> {
-    Optional<T> validate(Function<T, Optional<T>> validator);
-  }
-}
