+++ src/main/java/org/folio/marccat/integration/MarccatHelper.java
@@ -3,30 +3,33 @@ package org.folio.marccat.integration;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import org.folio.marccat.exception.DataAccessException;
 import org.folio.marccat.exception.SystemInternalFailureException;
+import org.folio.marccat.exception.UnableToCreateOrUpdateEntityException;
 import org.springframework.boot.jdbc.DataSourceBuilder;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 
 import javax.sql.DataSource;
 import java.sql.Connection;
 import java.sql.SQLException;
-import java.util.AbstractMap;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Properties;
+import java.util.*;
+import java.util.function.BooleanSupplier;
+import java.util.function.Function;
 import java.util.stream.StreamSupport;
 
 import static java.util.stream.Collectors.toMap;
 import static org.folio.marccat.config.Global.HCONFIGURATION;
 
 /**
- * Helper functions used within the marccat module.
+ * Helper functions used within the cataloging module.
  * Specifically, this class was originally thought as a supertype layer of each resource implementor; later, it has
  * been converted in this way (i.e. a collection of static methods) because https://issues.folio.org/browse/RMB-95
  *
- * @author cchiama
+ * @author agazzarini
  * @since 1.0
  */
 public abstract class MarccatHelper {
-  private static final Properties DEFAULT_VALUES = new Properties();
+  private final static Properties DEFAULT_VALUES = new Properties();
   private final static Map<String, DataSource> DATASOURCES = new HashMap<>();
 
   static {
@@ -54,7 +57,70 @@ public abstract class MarccatHelper {
   }
 
   /**
-   * Provides a unified approach (within the marccat module) for wrapping an existing blocking flow.
+   * Executes a POST request.
+   *
+   * @param adapter           the bridge that carries on the existing logic.
+   * @param tenant            the tenant associated with the current request.
+   * @param configurator      the configuration client.
+   * @param validator         a validator function for the entity associated with this resource.
+   * @param configurationSets the requested configuration attributes sets.
+   */
+  public static <T> ResponseEntity<T> doPost(
+    final PieceOfExistingLogicAdapter<T> adapter,
+    final String tenant,
+    final Configuration configurator,
+    final BooleanSupplier validator,
+    final String... configurationSets) {
+    if (validator.getAsBoolean()) {
+      final T result = exec(adapter, tenant, configurator, configurationSets);
+      final HttpHeaders headers = new HttpHeaders();
+      headers.add(HttpHeaders.CONTENT_TYPE, "application/json");
+      return new ResponseEntity<>(result, headers, HttpStatus.CREATED);
+    } else {
+      throw new UnableToCreateOrUpdateEntityException();
+    }
+  }
+
+  /**
+   * Executes a PUT request.
+   *
+   * @param adapter           the bridge that carries on the existing logic.
+   * @param tenant            the tenant associated with the current request.
+   * @param configurator      the configuration client.
+   * @param validator         a validator function for the entity associated with this resource.
+   * @param configurationSets the requested configuration attributes sets.
+   */
+  public static <T> void doPut(
+    final PieceOfExistingLogicAdapter<T> adapter,
+    final String tenant,
+    final Configuration configurator,
+    final BooleanSupplier validator,
+    final String... configurationSets) {
+    if (validator.getAsBoolean()) {
+      exec(adapter, tenant, configurator, configurationSets);
+    } else {
+      throw new UnableToCreateOrUpdateEntityException();
+    }
+  }
+
+  /**
+   * Executes a DELETE request.
+   *
+   * @param adapter           the bridge that carries on the existing logic.
+   * @param tenant            the tenant associated with the current request.
+   * @param configurator      the configuration client.
+   * @param configurationSets the requested configuration attributes sets.
+   */
+  public static <T> void doDelete(
+    final PieceOfExistingLogicAdapter<T> adapter,
+    final String tenant,
+    final Configuration configurator,
+    final String... configurationSets) {
+    exec(adapter, tenant, configurator, configurationSets);
+  }
+
+  /**
+   * Provides a unified approach (within the cataloging module) for wrapping an existing blocking flow.
    *
    * @param adapter           the bridge that carries on the existing logic.
    * @param configurationSets the configurationSets required by the current service.
@@ -89,8 +155,7 @@ public abstract class MarccatHelper {
    * @return a dedicated configuration for the current service.
    */
   private static Map<String, String> configuration(final ObjectNode value) {
-    return StreamSupport
-      .stream(value.withArray("configs").spliterator(), false)
+    return StreamSupport.stream(value.withArray("configs").spliterator(), false)
       .filter(node -> !"datasource".equals(node.get("configName").asText()))
       .map(node -> new AbstractMap.SimpleEntry<>(node.get("code").asText(), node.get("value").asText()))
       .collect(toMap(AbstractMap.SimpleEntry::getKey, AbstractMap.SimpleEntry::getValue));
@@ -121,8 +186,16 @@ public abstract class MarccatHelper {
       .create()
       .username(config.get("user"))
       .password(config.get("password"))
-      .url("jdbc:postgresql://192.168.0.158:5433/folio_marccat_sv2")
+      .url(config.get("url"))
       .build();
   }
 
+  /**
+   * A simple definition of a validation interface.
+   *
+   * @param <T> the kind of object that needs to be validated.
+   */
+  interface Valid<T> {
+    Optional<T> validate(Function<T, Optional<T>> validator);
+  }
 }
