+++ src/main/java/org/folio/marccat/resources/BrowseAPI.java
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.function.Function;
 
 import static java.util.stream.Collectors.toList;
-import static org.folio.marccat.integration.CatalogingHelper.doGet;
+import static org.folio.marccat.integration.MarccatHelper.doGet;
 
 /**
  * Headings RESTful APIs.
