+++ src/main/java/org/folio/marccat/resources/FieldAPI.java
@@ -19,7 +19,7 @@ import java.util.Map;
 import static java.util.Arrays.asList;
 import static java.util.Arrays.stream;
 import static java.util.stream.Collectors.toList;
-import static org.folio.marccat.integration.CatalogingHelper.doGet;
+import static org.folio.marccat.integration.MarccatHelper.doGet;
 
 /**
  * FieldsAPI RestFul service to manage fields (tag).
