+++ src/main/java/org/folio/marccat/resources/FixedFieldCodesGroupAPI.java
@@ -21,7 +21,7 @@ import java.util.function.Function;
 
 import static java.util.Optional.ofNullable;
 import static java.util.stream.Collectors.toList;
-import static org.folio.marccat.integration.CatalogingHelper.doGet;
+import static org.folio.marccat.integration.MarccatHelper.doGet;
 
 /**
  * Fixed-Field Codes Groups RESTful APIs.
